Added routing to existing main maps
Currently, if you go to an existing main map and upload a new image, it just adds the image to the main_maps folder but also keeps the old one, it needs to also delete the old one if you upload a new one in the same page (i think). Then, now we also need to link the pages for existing campaigns to maps and just change the header aswell.